Add files via upload
In this project I compared data from the Google Play Store and iOS App Store to try and determine an app profile that would be profitable in both markets. I concluded that by taking a book and including special features like daily quotes and a audio version would be popular because both markets were already full of libraries.